Sync deps: libcat
* libcat/libcat@62c970f2d8f8ac04e88d35b751930a4379f39f66

+++ ext/deps/libcat/src/cat_memory.c
@@ -81,7 +81,7 @@ CAT_API char *cat_strndup_function(const char *string, size_t length)
     return cat_strndup(string ,length);
 }
 
-CAT_API size_t cat_getpagesize(void)
+CAT_API int cat_getpagesize(void)
 {
     static int pagesize = 0;
 
@@ -103,7 +103,7 @@ CAT_API size_t cat_getpagesize(void)
         }
     }
 
-    return (size_t) pagesize;
+    return (int) pagesize;
 }
 
 CAT_API void *cat_getpageof(const void *p)
